requirements.txt
openai>=1.30.0,<2.0.0
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
markdown>=3.6
pytest>=8.0.0
pytest-mock>=3.12.0
